ML-COVID-Detection
This repository contains code and resources for a machine learning project focused on COVID-19 detection. The project includes data preprocessing, model training, and evaluation scripts. It also provides a Jupyter notebook with detailed explanations of the analysis and results. The goal of this project is to develop a predictive model for COVID-19 based on various clinical features and symptoms.