Customer-Data-Analysis
This repository contains code and resources for an analysis of customer data. The project includes data preprocessing steps such as dropping unnecessary columns and encoding categorical variables using scikit-learn. Specifically, it demonstrates the use of OrdinalEncoder for categorical features ('Education' and 'Income') with predefined categories and LabelEncoder for encoding the target variable. The goal of this project is to prepare the data for machine learning tasks and showcases the encoding process for categorical variables in a real-world dataset.