random_imputation_data_preprocessing
The random_imputation_data_preprocessing.py script exemplifies advanced data preprocessing techniques for handling missing values in datasets. This script is designed to be a versatile tool for data analysts and data scientists, showcasing the methodology of random imputation to address missing data.

Key Features:

Dataset Handling: The script demonstrates how to load datasets efficiently using the Pandas library, focusing on relevant columns such as 'Age,' 'Fare,' 'Survived' (in a Titanic dataset), 'GarageQual,' 'FireplaceQu,' and 'SalePrice' (in a housing dataset).

Missing Data Analysis: It calculates and visualizes the percentage of missing data in each column, providing valuable insights into the dataset's quality.

Train-Test Split: The script performs a train-test split, an essential step for machine learning projects, ensuring the separation of data for training and evaluation.

Random Imputation: A core feature of this script, random imputation, is applied to the 'Age' column in the Titanic dataset and the 'GarageQual' and 'FireplaceQu' columns in the housing dataset. Missing values are imputed with random values sampled from the non-null values within the same column.

Variance Analysis: The script compares the variance of original and imputed data to assess the impact of imputation on data variability.

Covariance Analysis: It computes and presents the covariance matrix between relevant columns, aiding in understanding relationships within the data.

Data Visualization: The script uses Seaborn and Matplotlib for data visualization, providing insightful visual representations of original and imputed data distributions.

This comprehensive Python script serves as a valuable resource for data preprocessing tasks, especially when dealing with datasets containing missing values. It showcases the power and flexibility of random imputation as one of the techniques for handling incomplete data. Users can adapt and extend this script to suit their specific data preprocessing needs in various projects.