Missing_Values_Imputation_and_Data_Exploration
This Python script demonstrates the process of handling missing values in a dataset and exploring the impact of imputation on a target variable. Specifically, the script focuses on two columns, 'GarageQual' and 'FireplaceQu', along with the 'SalePrice' variable. Here's a summary of what the script accomplishes:

Loads a dataset from a CSV file and selects relevant columns ('GarageQual', 'FireplaceQu', 'SalePrice').

Calculates and visualizes the percentage of missing values in the dataset.

Analyzes the distribution of 'GarageQual' values through bar charts and imputes missing values with the most frequent value ('TA').

Visualizes the effect of imputation on the 'SalePrice' distribution, both before and after imputation, using Kernel Density Estimation (KDE) plots.

Performs a similar analysis for the 'FireplaceQu' column, imputing missing values with the most frequent value ('Gd').

Explores the impact of 'FireplaceQu' imputation on the 'SalePrice' distribution.

Splits the data into training and testing sets using Scikit-Learn's train_test_split.

Imputes missing values in both the training and testing sets using Scikit-Learn's SimpleImputer with the most frequent value strategy.

Displays the statistics representing the most frequent values used for imputation.

This script serves as an illustrative example of how to handle missing data in specific columns and assess the impact of imputation on a target variable. It can be a valuable resource for data preprocessing tasks in machine learning projects involving datasets with missing values and exploring the consequences of different imputation strategies on the data distribution.